+++ serverless/src/handlers/createOrder.ts
@@ -1,7 +1,7 @@
-import { createOrder } from "../data"
-import { CreatePresignedUrlResponse } from "./createPresignedUrl"
+import { createOrder } from "../services/prodigi"
+// import { EventBridgeEvent } from "aws-lambda"
 
-export const handler = async (event: CreatePresignedUrlResponse) => {
+export const handler = async (event: {}) => {
     console.log(event)
     await createOrder()
 }
