+++ serverless/src/handlers/sendEmail.ts
@@ -1,26 +1,24 @@
+import { SQSEvent, SQSRecord } from "aws-lambda"
 import Stripe from "stripe"
 
-import { sendEmail } from "../data"
+import { sendEmail } from "../services/ses"
 
-export const handler = async (event: {
-    detail: Stripe.Event
-}): Promise<void> => {
+export const handler = async (event: SQSEvent): Promise<void> => {
     try {
-        // if (stripeEvent.type !== 'checkout.session.completed') {
-        //     // Do something
-        // }
+        for (const record of event.Records as SQSRecord[]) {
+            const body = JSON.parse(record.body)
+            const stripeEvent = body.detail as Stripe.Event
+            const sessionId = stripeEvent.data.object.id
+            const { error } = await sendEmail({ sessionId })
 
-        const sessionId = event.detail.data.object.id
-        const { error } = await sendEmail({ sessionId })
+            if (error) {
+                console.error(error)
 
-        if (error) {
-            console.error(error)
-
-            throw new Error("Could not process SQS event record")
+                throw new Error("Could not process SQS event record")
+            }
         }
     } catch (error) {
         console.error(error)
-
         throw error
     }
 }
